% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnSubsetChisq.R
\name{EnSubsetChisq}
\alias{EnSubsetChisq}
\title{Ensemble Subset Chi-squared test}
\usage{
EnSubsetChisq(
  X,
  Sigma,
  B,
  m = NULL,
  is.Sigma.identity = FALSE,
  is.pvals.path = FALSE
)
}
\arguments{
\item{X}{a numeric vector of z-scores.}

\item{Sigma}{the correlation matrix of the z-scores.}

\item{B}{the number of base tests.}

\item{m}{the length of the subset of X. If not provided, m is set to be floor(sqrt(length(X))).}

\item{is.Sigma.identity}{logical. If \code{TRUE}, \code{Sigma} will be set to be an identity matrix and the implementation would be more efficient.}

\item{is.pvals.path}{logical. If \code{TRUE}, the p-value path of the ensemble test as \emph{B} increases will be provided in the output and can be used to draw the ensemble p-value path plot (See examples).}
}
\value{
A list with components:
\describe{
\item{\code{pval.ensemble.test}}{The p-value of the ensemble test}
\item{\code{pval.base.test}}{The p-values of the individual base tests}
\item{\code{pval.path.ensemble}}{The p-value path of the ensemble test when \emph{is.pvals.path == TRUE}.}
}
}
\description{
Ensemble Subset Chi-squared test
}
\examples{
#### Example 1, Sigma is an identity matrix
set.seed(1234); p = 20
X = rnorm(p)
X[1:3] = X[1:3]+3.5   ## z-scores
res = EnSubsetChisq(X,Sigma = diag(p),B=1000,is.Sigma.identity = T, is.pvals.path = T)
res$pval.ensemble.test   ## the ensemble p-value

pvals.path = res$pval.path.ensemble   ## p-value path plot
plot(c(1:length(pvals.path)), -log10(pvals.path),xlab = "Number of base tests",ylab = "-log10(p-value)",type = "l")

#### Example 2, Sigma is not an identity matrix
library(MASS)
set.seed(1234); p = 20
Sigma = matrix(0.1,nrow = p, ncol = p); diag(Sigma) = 1;
X = mvrnorm(1,rep(0,p),Sigma)
X[1:3] = X[1:3]+3.5   ## z-scores
res = EnSubsetChisq(X,Sigma =Sigma,B=1000, is.pvals.path = T)
res$pval.ensemble.test  ## the ensemble p-value

pvals.path = res$pval.path.ensemble  ## p-value path plot
plot(c(1:length(pvals.path)), -log10(pvals.path),xlab = "Number of base tests",ylab = "-log10(p-value)",type = "l")

}
\references{
Liu, Y., Liu, Z., and Lin, X. (2024+) Ensemble methods for testing a global null.
\emph{Journal of the Royal Statistical Society: Series B (Statistical Methodology)}. Accepted.
}
\author{
Yaowu Liu
}
